requirements.txt
absl-py==1.1.0
attrs==21.4.0
cycler==0.11.0
fonttools==4.33.3
kiwisolver==1.4.3
matplotlib==3.5.2
mediapipe-silicon==0.8.9; sys.platform == 'darwin' and platform.machine == 'arm64'
mediapipe==0.8.9; sys.platform != 'darwin' or platform.machine != 'arm64'
numpy==1.22.4
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.1.1
protobuf==4.21.1
pyparsing==3.0.9
python-dateutil==2.8.2
six==1.16.0
